+++ src/test/extension.test.ts
@@ -37,10 +37,7 @@ suite('Extension Test Suite - txt output', () => {
 		// MEMO : command 後、inputbox へ値をセットすることは難しいらいため、UI操作ではなくControllerを操作
 		// vscode.commands.executeCommand('extension.grepResult2File');
 		const controller = new GrepController();
-		controller.doActionWithParam('lo');
-
-		// Wait for the process to complete
-		await (async () => {return new Promise(resolve => setTimeout(resolve, 5000));})();
+		await controller.doActionWithParam('lo');
 
 		// Get actual result
 		const actualValue = (await vscode.window.showTextDocument(vscode.Uri.file(inputFilePath))).document.getText();
@@ -74,10 +71,7 @@ suite('Extension Test Suite - txt output', () => {
 		// MEMO : command 後、inputbox へ値をセットすることは難しいらいため、UI操作ではなくControllerを操作
 		// vscode.commands.executeCommand('extension.grepResult2File');
 		const controller = new GrepController();
-		controller.doActionWithParam('re/lo.*it/');
-
-		// Wait for the process to complete
-		await (async () => {return new Promise(resolve => setTimeout(resolve, 5000));})();
+		await controller.doActionWithParam('re/lo.*it/');
 
 		// Get actual result
 		const actualValue = (await vscode.window.showTextDocument(vscode.Uri.file(inputFilePath))).document.getText();
@@ -117,10 +111,7 @@ suite('Extension Test Suite - tsv output', () => {
 		// MEMO : command 後、inputbox へ値をセットすることは難しいらいため、UI操作ではなくControllerを操作
 		// vscode.commands.executeCommand('extension.grepResult2File');
 		const controller = new GrepController();
-		controller.doActionWithParam('lo');
-
-		// Wait for the process to complete
-		await (async () => {return new Promise(resolve => setTimeout(resolve, 5000));})();
+		await controller.doActionWithParam('lo');
 
 		// Get actual result
 		const actualValue = (await vscode.window.showTextDocument(vscode.Uri.file(inputFilePath))).document.getText();
@@ -154,10 +145,7 @@ suite('Extension Test Suite - tsv output', () => {
 		// MEMO : command 後、inputbox へ値をセットすることは難しいらいため、UI操作ではなくControllerを操作
 		// vscode.commands.executeCommand('extension.grepResult2File');
 		const controller = new GrepController();
-		controller.doActionWithParam('re/lo.*it/');
-
-		// Wait for the process to complete
-		await (async () => {return new Promise(resolve => setTimeout(resolve, 5000));})();
+		await controller.doActionWithParam('re/lo.*it/');
 
 		// Get actual result
 		const actualValue = (await vscode.window.showTextDocument(vscode.Uri.file(inputFilePath))).document.getText();
@@ -197,10 +185,7 @@ suite('Extension Test Suite - csv output', () => {
 		// MEMO : command 後、inputbox へ値をセットすることは難しいらいため、UI操作ではなくControllerを操作
 		// vscode.commands.executeCommand('extension.grepResult2File');
 		const controller = new GrepController();
-		controller.doActionWithParam('lo');
-
-		// Wait for the process to complete
-		await (async () => {return new Promise(resolve => setTimeout(resolve, 5000));})();
+		await controller.doActionWithParam('lo');
 
 		// Get actual result
 		const actualValue = (await vscode.window.showTextDocument(vscode.Uri.file(inputFilePath))).document.getText();
@@ -234,10 +219,7 @@ suite('Extension Test Suite - csv output', () => {
 		// MEMO : command 後、inputbox へ値をセットすることは難しいらいため、UI操作ではなくControllerを操作
 		// vscode.commands.executeCommand('extension.grepResult2File');
 		const controller = new GrepController();
-		controller.doActionWithParam('re/lo.*it/');
-
-		// Wait for the process to complete
-		await (async () => {return new Promise(resolve => setTimeout(resolve, 5000));})();
+		await controller.doActionWithParam('re/lo.*it/');
 
 		// Get actual result
 		const actualValue = (await vscode.window.showTextDocument(vscode.Uri.file(inputFilePath))).document.getText();
@@ -246,6 +228,78 @@ suite('Extension Test Suite - csv output', () => {
 		// Assert
 		// ---------------------------
 		assert.equal(expectedValue, actualValue);
+	});
+
+});
+
+suite('Extension Test Suite - json output', () => {
+	vscode.window.showInformationMessage('Start all tests.');
+
+	const inputFilePath = path.join(inputFolderPath, 'grep2File.g2f.json');
+
+	test('Grep word by default mode - lo', async () => {
+
+		// ---------------------------
+		// Arrange
+		// ---------------------------
+		// Clear Editor Content
+		(await vscode.window.showTextDocument(vscode.Uri.file(inputFilePath))).edit(editBuilder => {
+			editBuilder.delete(new vscode.Range(new vscode.Position(0,0), new vscode.Position(1000,0)));
+		});
+		// Setting : json
+		await vscode.workspace.getConfiguration().update('grep2file.outputContentFormat', 'json', vscode.ConfigurationTarget.Global);
+		// Get expected result
+		const expectedFilePath = path.join(expectedFolderPath, 'grep2File.g2f.json');
+		const expectedValue = fs.readFileSync(expectedFilePath, 'utf-8');
+
+		// ---------------------------
+		// Action
+		// ---------------------------
+		// Execute command
+		const controller = new GrepController();
+		await controller.doActionWithParam('lo');
+
+		// Get actual result
+		const actualValue = (await vscode.window.showTextDocument(vscode.Uri.file(inputFilePath))).document.getText();
+
+		// ---------------------------
+		// Assert
+		// ---------------------------
+		assert.equal(expectedValue, actualValue);
+		assert.doesNotThrow(() => JSON.parse(actualValue));
+
+	});
+
+	test('Grep word by regexp mode - re/lo.*it/', async () => {
+
+		// ---------------------------
+		// Arrange
+		// ---------------------------
+		// Clear Editor Content
+		(await vscode.window.showTextDocument(vscode.Uri.file(inputFilePath))).edit(editBuilder => {
+			editBuilder.delete(new vscode.Range(new vscode.Position(0,0), new vscode.Position(1000,0)));
+		});
+		// Setting : json
+		await vscode.workspace.getConfiguration().update('grep2file.outputContentFormat', 'json', vscode.ConfigurationTarget.Global);
+		// Get expected result
+		const expectedFilePath = path.join(expectedFolderPath, 'regexp_grep2File.g2f.json');
+		const expectedValue = fs.readFileSync(expectedFilePath, 'utf-8');
+
+		// ---------------------------
+		// Action
+		// ---------------------------
+		// Execute command
+		const controller = new GrepController();
+		await controller.doActionWithParam('re/lo.*it/');
+
+		// Get actual result
+		const actualValue = (await vscode.window.showTextDocument(vscode.Uri.file(inputFilePath))).document.getText();
+
+		// ---------------------------
+		// Assert
+		// ---------------------------
+		assert.equal(expectedValue, actualValue);
+		assert.doesNotThrow(() => JSON.parse(actualValue));
 	});
 
 });